+++ README.md
@@ -8,3 +8,7 @@ cmsenv
 git cms-init
 git clone git@github.com:dildick/dt_analysis_code.git
 </PRE>
+
+Running the signal analysis
+
+root -l -q -b analyzeSignal.C++
